backend/requirements.txt
streamlit
sentence-transformers
faiss-cpu
transformers
torch
pydantic # Good to keep as a dependency of other libraries
fastapi
uvicorn[standard]
requests